Implemented Chess\Variant\RandomStartPieces
